refactor: split File menu overlay out of views_overlays
views_overlays.rs crossed the 1,000-line ceiling (1,056) when the
Recent Projects card landed. The File menu, its submenu-alignment
constants, and their test move to views_file_menu.rs (275 lines),
bringing views_overlays.rs to 809.

Also fixes the alignment invariant the submenu arithmetic assumes:
recent_btn and settings_btn were the only menu rows missing
.height(FILE_MENU_ITEM_HEIGHT), so the Recent row rendered taller
than the 32px the offset math is built on.
