+++ crates/vibez-ui/src/app/mod.rs
@@ -185,6 +185,7 @@ mod views_clip_swing;
 mod views_close_confirm;
 mod views_detail;
 mod views_devices;
+mod views_file_menu;
 mod views_mixer;
 mod views_overlays;
 mod views_perform;
